Removed boilerplate for client and factory combo
