+++ src/test/kotlin/com/terraformation/backend/support/AtlassianHttpClientExternalTest.kt
@@ -3,7 +3,6 @@ package com.terraformation.backend.support
 import com.terraformation.backend.config.TerrawareServerConfig
 import com.terraformation.backend.support.atlassian.AtlassianHttpClient
 import com.terraformation.backend.support.atlassian.SupportRequestType
-import com.terraformation.backend.support.atlassian.resource.AtlassianResourceFactory
 import java.net.URI
 import org.junit.Assume
 import org.junit.jupiter.api.AfterEach
@@ -14,7 +13,6 @@ import org.springframework.boot.context.properties.EnableConfigurationProperties
 
 @EnableConfigurationProperties(TerrawareServerConfig::class)
 class AtlassianHttpClientExternalTest {
-  private lateinit var factory: AtlassianResourceFactory
   private lateinit var client: AtlassianHttpClient
   private val createdIssueIds: MutableList<String> = mutableListOf()
 
@@ -47,28 +45,22 @@ class AtlassianHttpClientExternalTest {
                     apiClientUsernamePrefix = "test"))
 
     client = AtlassianHttpClient(config)
-    factory = AtlassianResourceFactory(config)
   }
 
   @AfterEach
   fun deleteCreatedIssues() {
-    createdIssueIds.forEach {
-      val request = factory.deleteIssue(it)
-      client.makeRequest(request)
-    }
+    createdIssueIds.forEach { client.deleteIssue(it) }
     createdIssueIds.clear()
   }
 
   @Test
   fun `create new issue`() {
-    val request =
-        factory.createServiceDeskRequest(
+    val response =
+        client.createServiceDeskRequest(
             description = "Description",
             summary = "Summary",
             reporter = "testuser@example.com",
             requestType = SupportRequestType.FEATURE_REQUEST)
-
-    val response = client.makeRequest(request)
 
     assertNotNull(response)
     createdIssueIds.addLast(response.issueId)
